Updated utility.js by adding a new function getDueDateString()! Updated AdminAddAssessmentTask.js by importing the new function getDueDateString()!
